+++ src/public/src/components/pages/dashboard.js
@@ -1,7 +1,7 @@
 import React from "react";
-import NavBar from "./dash/navbar";
-import HomeDrawer from "./dash/homeDrawer";
-import Links from "./dash/links";
+import NavBar from "../dash/navbar";
+import HomeDrawer from "../dash/homeDrawer";
+import Links from "../dash/links";
 
 const Dashboard = () => {
   return (
